+++ components/restreamer/client/src/components/Console.svelte
@@ -1,4 +1,7 @@
 <script lang="js">
+  import Fa from 'svelte-fa';
+  import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
+
   import { mutation, subscribe } from 'svelte-apollo';
   import { ConsoleLog } from '../../api/dashboard.graphql';
   import { ConsoleClear } from '../../api/dashboard.graphql';
@@ -39,19 +42,25 @@
       >
       {#if errorCount}
         <span class="console-icon" title="Number of errors">
-          <i class="fa fa-exclamation-triangle console-error" />{errorCount}
+          <span class="icon-exclamation-triangle console-error"
+            ><Fa icon={faExclamationTriangle} /></span
+          >{errorCount}
         </span>
       {/if}
 
       {#if warningCount}
         <span class="console-icon" title="Number of warning messages">
-          <i class="fa fa-exclamation-triangle console-warning" />{warningCount}
+          <span class="icon-exclamation-triangle console-warning"
+            ><Fa icon={faExclamationTriangle} /></span
+          >{warningCount}
         </span>
       {/if}
 
       {#if infoCount}
         <span class="console-icon" title="Number of info messages">
-          <i class="fa fa-exclamation-triangle" />{infoCount}
+          <span class="icon-exclamation-triangle"
+            ><Fa icon={faExclamationTriangle} /></span
+          >{infoCount}
         </span>
       {/if}
 
@@ -112,7 +121,7 @@
   .console-icon
     padding: 8px
 
-  .fa-exclamation-triangle
+  .icon-exclamation-triangle
     display: inline-block
     margin-right: 4px