+++ components/restreamer/client/src/components/Mixin.svelte
@@ -1,4 +1,9 @@
 <script lang="js">
+  import Fa from 'svelte-fa';
+  import { faWaveSquare } from '@fortawesome/free-solid-svg-icons';
+  import { faClock } from '@fortawesome/free-regular-svg-icons';
+  import { faLink } from '@fortawesome/free-solid-svg-icons';
+
   import { mutation } from 'svelte-apollo';
   import { showError } from '../utils/util';
   import Volume from './common/Volume.svelte';
@@ -78,7 +83,9 @@
 
 <template>
   <div class="mixin">
-    <i class="fas fa-wave-square" title="Mixed audio" />
+    <span class="icon-wave-square" title="Mixed audio"
+      ><Fa icon={faWaveSquare} /></span
+    >
     <Url url={hideIdentity(value.src)} />
     <Volume
       volume={value.volume}
@@ -90,7 +97,7 @@
       title="Mixed audio"
     />
     <div class="mixin-options">
-      <i class="far fa-clock" title="Delay" />
+      <span class="icon-clock"><Fa icon={faClock} /></span>
       <span>Delay</span>
       <input
         class="uk-input"
@@ -102,7 +109,7 @@
         title="Delay"
       />
       <span>s</span>
-      <i class="fas fa-link" title="Sidechain" />
+      <span class="icon-link"><Fa icon={faLink} /></span>
       <span>Sidechain</span>
       <input
         class="uk-checkbox"
@@ -117,7 +124,7 @@
 </template>
 
 <style lang="stylus">
-  .fa-wave-square, .fa-clock, .fa-link
+  .icon-wave-square, .icon-clock, .icon-link
     font-size: 10px
     color: #d9d9d9
 
@@ -136,7 +143,7 @@
       margin-top: -2px
       text-align: right
 
-    .fa-link
+    .icon-link
       margin-left: 15px
 
     .uk-checkbox