+++ components/restreamer/client/src/components/common/StatusFilter.svelte
@@ -1,5 +1,8 @@
 <script lang="js">
+  import Fa from 'svelte-fa';
+  import { faInfoCircle } from '@fortawesome/free-solid-svg-icons/faInfoCircle';
   import { STREAM_ERROR, STREAM_WARNING } from '../../utils/constants';
+  import ToggleButton from './ToggleButton.svelte';
 
   export let count;
   export let active;
@@ -7,68 +10,45 @@
   export let disabled;
   export let title;
   export let handleClick = () => {};
+
+  $: iconClass = getClass(status);
+
+  function getClass(s) {
+    let cls = 'info-icon';
+
+    if (s === STREAM_ERROR) {
+      cls += ' streams-errors';
+    } else if (s === STREAM_WARNING) {
+      cls += ' streams-warnings';
+    }
+
+    return cls;
+  }
 </script>
 
 <template>
-  <!-- svelte-ignore a11y-click-events-have-key-events -->
-  <div
-    class="status-filter"
-    on:click={(e) => {
-      if (disabled) {
-        return;
-      }
-
-      handleClick(e);
-    }}
-  >
+  <ToggleButton {handleClick} {disabled} {active}>
     <div
       title={title ? title : status}
-      class="content"
-      class:active
-      class:disabled
       class:online={status === 'ONLINE'}
       class:offline={status === 'OFFLINE'}
       class:initializing={status === 'INITIALIZING'}
       class:unstable={status === 'UNSTABLE'}
     >
       {#if [STREAM_ERROR, STREAM_WARNING].includes(status)}
-        <i
-          class:streams-errors={status === STREAM_ERROR}
-          class:streams-warnings={status === STREAM_WARNING}
-          class="fa fa-info-circle info-icon"
-        />
+        <Fa class={iconClass} icon={faInfoCircle} />
       {:else}
         <span class="circle" />
       {/if}
       {count}
     </div>
-  </div>
+  </ToggleButton>
 </template>
 
 <style lang="stylus">
-  .status-filter
-    min-width: 32px
-    display: inline-flex
-    .content
-      width: 100%
-      text-align: center
-      margin-right: 2px
-      background-color: inherit
-      padding: 1px 4px
-      border-radius: 2px
-      outline: none
-      &.active
-        background-color: #cecece
-      &.disabled
-        &:hover
-          cursor: not-allowed
-      &:hover
-        background-color: #bdbdbd
-        cursor: pointer
-
-  .streams-errors
+  :global(.streams-errors)
     color: var(--danger-color)
-  .streams-warnings
+  :global(.streams-warnings)
     color: var(--warning-color)
 
 </style>