+++ components/restreamer/client/src/utils/util.ts
@@ -4,7 +4,8 @@ import { SubscriptionClient } from 'subscriptions-transport-ws';
 import { ApolloClient } from '@apollo/client/core';
 import { WebSocketLink } from '@apollo/client/link/ws';
 import { InMemoryCache } from '@apollo/client/cache';
-
+import isEqual from 'lodash/isEqual';
+import take from 'lodash/take';
 /**
  * Displays an error UI popup with the given error `message`.
  *
@@ -177,4 +178,8 @@ export const getFolderIdFromGDrive = (id) => {
   }
 
   return id;
+};
+
+export const isArrayStartWithAnother = (arr1: [], arr2: []) => {
+  return isEqual(arr1, take(arr2, arr1.length));
 };